refactor(dtos): use barrel export and improve dtos related with chat
