+++ backend/src/chats/dto/createChat.dto.ts
@@ -1,7 +1,8 @@
-import { ArrayMinSize, IsArray } from 'class-validator';
+import { ArrayMinSize, IsArray, IsNotEmpty } from 'class-validator';
 
 export class CreateChatDto {
   @IsArray()
   @ArrayMinSize(1)
+  @IsNotEmpty()
   userIds: number[];
 }
