+++ backend/src/chats/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './createChat.dto';
+export * from './updateChat.dto';
+export * from './chatWithName.dto';
